+++ migrations/2023-02-05-221140_marker_images/up.sql
@@ -4,9 +4,9 @@
 
 CREATE TABLE IF NOT EXISTS public.marker_images
 (
-    id SERIAL NOT NULL,
+    id BIGSERIAL NOT NULL,
     path text NOT NULL,
-    refers_to integer NOT NULL,
+    refers_to BIGINT NOT NULL,
     CONSTRAINT marker_images_id_pkey PRIMARY KEY (id),
     CONSTRAINT marker_images_refers_to_fkey FOREIGN KEY (refers_to)
         REFERENCES public.markers (id) MATCH SIMPLE
